+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Unreleased
 
 ### Added
+- TEST: mf2 extension is loaded
+- Minimal PHP extension config
 - This CHANGELOG.md to track what's happening with this software.
 - An altogether too brief [README.md](README.md).
 - Licensed (MIT) this work, see [LICENSE.md](LICENSE.md) for details.